requirements.txt
flask~=2.1.1
flask-Cors~=3.0.8
flask-sqlalchemy~=2.5.1
SQLAlchemy~=1.4
Werkzeug~=2.1.1
Flask-Migrate~=2.5.3
gunicorn~=20.1.0
python-dotenv~=0.15.0
jinja2~=3.1.1
pytest~=7.0.1
requests~=2.27.1
flask-login
nanoid~=2.0.0
sendgrid
python-dateutil
celery~=5.2
pydantic
sqlalchemy_utils
redis
